# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "Body_upload_wsr_api_v1_wsr_uploads_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_wsr_api_v1_wsr_uploads_post",
        "type": "object"
      },
...
      "EvidenceReference": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "dependency_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependency Description"
          },
          "predecessor_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predecessor Names"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "resource_assignments": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Assignments"
          },
          "task_or_milestone_name": {
            "title": "Task Or Milestone Name",
            "type": "string"
          }
        },
        "required": [
          "task_or_milestone_name"
        ],
        "title": "EvidenceReference",
        "type": "object"
      },
      "FileRecord": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "extracted_char_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Char Count"
          },
          "extracted_text_available": {
            "default": false,
            "title": "Extracted Text Available",
            "type": "boolean"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_accessed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Accessed At"
          },
          "module": {
            "anyOf": [
              {
                "enum": [
                  "sow",
                  "wsr",
                  "retrospective",
                  "plan"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "plan_available": {
            "default": false,
            "title": "Plan Available",
            "type": "boolean"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "filename",
          "content_type",
          "size",
          "created_at"
        ],
        "title": "FileRecord",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProcessingResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "module": {
            "enum": [
              "sow",
              "wsr",
              "retrospective",
              "plan"
            ],
            "title": "Module",
            "type": "string"
          },
          "request_handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Handle"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "module",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProcessingResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WsrEvidenceResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "evidence_references": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence References",
            "type": "array"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "review_status": {
            "enum": [
              "pending",
              "kept",
              "edited",
              "removed"
            ],
            "title": "Review Status",
            "type": "string"
          },
          "section": {
            "title": "Section",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "content",
          "section",
          "review_status",
          "evidence_references"
        ],
        "title": "WsrEvidenceResponse",
        "type": "object"
      },
      "WsrItemDecision": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "decision": {
            "enum": [
              "kept",
              "edited",
              "removed"
            ],
            "title": "Decision",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "WsrItemDecision",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local MVP API. AUTH_MODE=disabled skips sign-in. AUTH_MODE=required issues HttpOnly session cookies. Public routes are versioned under /api/v1. ProcessingResponse.status is queued|running|succeeded|failed (blueprint state processing|complete|failed). See docs/OPENAPI_FREEZE.md.",
    "title": "Project Automation API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/wsr/requests/{handle}": {
      "get": {
        "operationId": "get_wsr_request_api_v1_wsr_requests__handle__get",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Wsr Request",
        "tags": [
          "wsr"
        ]
      }
    },
    "/api/v1/wsr/requests/{handle}/generate": {
      "post": {
        "operationId": "generate_wsr_request_api_v1_wsr_requests__handle__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Wsr Request",
        "tags": [
          "wsr"
        ]
      }
    },
    "/api/v1/wsr/requests/{handle}/items/{item_id}": {
      "patch": {
        "operationId": "review_wsr_item_api_v1_wsr_requests__handle__items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WsrItemDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Wsr Item",
        "tags": [
          "wsr"
        ]
      }
    },
    "/api/v1/wsr/requests/{handle}/items/{item_id}/evidence": {
      "get": {
        "operationId": "get_wsr_item_evidence_api_v1_wsr_requests__handle__items__item_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WsrEvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Wsr Item Evidence",
        "tags": [
          "wsr"
        ]
      }
    },
    "/api/v1/wsr/requests/{handle}/report": {
      "get": {
        "operationId": "download_wsr_report_api_v1_wsr_requests__handle__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Wsr Report",
        "tags": [
          "wsr"
        ]
      }
    },
    "/api/v1/wsr/uploads": {
      "post": {
        "operationId": "upload_wsr_api_v1_wsr_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_wsr_api_v1_wsr_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Wsr",
        "tags": [
          "wsr"
        ]
      }
    },
...
  }
}